requirements.txt
fastapi[standard]
uvicorn[standard]
pymongo
python-jose[cryptography]
passlib[bcrypt]
email-validator
python-multipart
google-genai